feat : distribute 조회 인터페이스 구현

+++ src/main/java/com/derekpark/balance/repository/DistributeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface DistributeRepository extends JpaRepository<Distribute, Long> {
+public interface DistributeRepository extends JpaRepository<Distribute, Integer> {
 
 }
